+++ api/src/team/application/organization-invitations/organization-invitation.controller.js
@@ -1,7 +1,6 @@
 import _ from 'lodash';
 
 import { MissingQueryParamError } from '../../../../lib/application/http-errors.js';
-import { usecases as libUsecases } from '../../../../lib/domain/usecases/index.js';
 import * as requestResponseUtils from '../../../shared/infrastructure/utils/request-response-utils.js';
 import { usecases } from '../../domain/usecases/index.js';
 import { organizationInvitationSerializer as organizationInvitationSerializer } from '../../infrastructure/serializers/jsonapi/organization-invitation.serializer.js';
@@ -15,7 +14,7 @@ import { serializer as scoOrganizationInvitationSerializer } from '../../infrast
  */
 const cancelOrganizationInvitation = async function (request, h) {
   const organizationInvitationId = request.params.organizationInvitationId;
-  await libUsecases.cancelOrganizationInvitation({ organizationInvitationId });
+  await usecases.cancelOrganizationInvitation({ organizationInvitationId });
   return h.response().code(204);
 };
